Controllers and Tables
Added first working table in Library/Index.cshtml that is a join of the
Song, Artist, Album, and Genre Table showing the list of songs we
currently have. Made some changes to create.sql to allow for unique
pairings that will prevent duplicates from ending up in the database.

Started working on Controllers and specifically the basic CRUD
operations within each table.

TODO: Determine if the models that were initially created will still
be needed. Determine the best way to make a UI to add things to tables